requirements.txt
grpcio==1.26.0
grpcio-tools==1.26.0
mysqlclient==1.4.6
protobuf==3.11.3
six==1.14.0
SQLAlchemy==1.3.13
